requirements.txt
# File: ~/VerbumTechnologies/semantic-tiles4-claude/requirements.txt
Flask==2.3.3
Flask-Cors==4.0.0
python-dotenv==1.0.0
PyPDF2==3.0.1
Werkzeug==2.3.7
sentence-transformers==2.2.2
transformers==4.35.0
torch==2.1.0
faiss-cpu==1.7.4
nltk==3.8.1
numpy==1.25.2
scikit-learn==1.3.0
gunicorn==21.2.0
uuid==1.30